+++ src/app/components/tasks-basic/task-form/task-basic-form.component.ts
@@ -44,6 +44,7 @@ import {LoadingOverlayService} from "@app/services/loading-overlay.service";
 import {LoggerService} from "@app/services/logger.service";
 import {UtilsService} from "@app/services/utils.service";
 import {magic} from "@environments/constants";
+import { compareNullableString } from '@app/utils/compare-nullable-string';
 
 /**
  * Component for managing basic tasks in the SITMUN application.
@@ -211,7 +212,7 @@ export class TaskBasicFormComponent extends BaseFormComponent<TaskProjection> {
       throw new Error(`UI tasks ${this.uiList} not retrieved`);
     }
 
-    taskGroups.sort((a, b) => a.name.localeCompare(b.name));
+    taskGroups.sort((a, b) => compareNullableString(a.name, b.name));
     this.taskGroupList = taskGroups;
   }